Restore recruiter panel thread configuration gating
