+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.9.5-d — Command registration consolidation (no behavior change) — 2025-10-22
+
+- Move recruiter command registration to `cogs/recruitment_recruiter.py`.
+- Remove legacy in-module registration; preserve existing UX and flags.
+- Continue honoring `PANEL_THREAD_MODE`/`PANEL_FIXED_THREAD_ID` when redirecting the panel thread.
+
 ## v0.9.5-a — Structure tidy (no behavior change) — 2025-10-22
 
 * Consolidate CoreOps to `modules/coreops`; remove `shared/coreops`.
